requirements.txt
fastapi
uvicorn
aiortc
opencv-python
ultralytics
supervision
huggingface_hub
Pillow
aiohttp

